% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-prscsx-sumstats.R
\name{prep_prscsx_sumstats}
\alias{prep_prscsx_sumstats}
\title{prep_prscsx_sumstats}
\usage{
prep_prscsx_sumstats(
  file,
  outfile,
  snp_col,
  effect_allele_col,
  other_allele_col,
  effect_col,
  pval_col,
  delim = "\\t"
)
}
\arguments{
\item{file}{filename containing GWAS summary statistics to prepare for PRS-CSx}

\item{outfile}{filename where prepared summary statistics should be written}

\item{snp_col}{unquoted column name containing SNP rsids}

\item{effect_allele_col}{unquoted column name containing effect alleles}

\item{other_allele_col}{unquoted column name containing non-effect alleles}

\item{effect_col}{unquoted column name containing effect sizes}

\item{pval_col}{unquoted column name containing p-values}

\item{delim}{delimiter used to define columns in the GWAS summary statistics file (default "\t")}
}
\value{
filename containing the prepared summary statistics
}
\description{
prep_prscsx_sumstats
}
